requirements.txt
torchtext
nltk
jieba